server/requirements.txt
absl-py==2.1.0
accelerate==0.30.0
adetailer @ git+https://github.com/Bing-su/adetailer.git@25e7509fe018de8aa063a5f1902598f5eda0c06c
annotated-types==0.7.0
attrs==24.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
compel==2.0.3
contourpy==1.2.1
cycler==0.12.1
diffusers==0.30.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.3.1
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.3
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
peft==0.12.0
pillow==10.3.0
protobuf==4.25.4
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
RealESRGAN @ git+https://github.com/sberbank-ai/Real-ESRGAN.git@362a0316878f41dbdfbb23657b450c3353de5acf
regex==2024.4.28
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.7
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.1
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.72
ultralytics-thop==2.0.0
urllib3==2.2.1
zipp==3.18.1
